+++ test/integration/helpers_test.go
@@ -13,10 +13,8 @@ import (
 	"github.com/samber/lo"
 	"github.com/stretchr/testify/require"
 	netv1 "k8s.io/api/networking/v1"
-	netv1beta1 "k8s.io/api/networking/v1beta1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/client-go/util/retry"
 	gatewayv1beta1 "sigs.k8s.io/gateway-api/apis/v1beta1"
 	gatewayclient "sigs.k8s.io/gateway-api/pkg/client/clientset/versioned"
@@ -393,32 +391,17 @@ func GetVerifyProgrammedConditionCallback(t *testing.T,
 
 // setIngressClassNameWithRetry changes Ingress.Spec.IngressClassName to specified value
 // and retries if update conflict happens.
-func setIngressClassNameWithRetry(ctx context.Context, namespace string, obj runtime.Object, ingressClassName *string) error {
-	switch ingress := obj.(type) {
-	case *netv1.Ingress:
-		ingressClient := env.Cluster().Client().NetworkingV1().Ingresses(namespace)
-		return retry.RetryOnConflict(retry.DefaultRetry, func() error {
-			ingress, err := ingressClient.Get(ctx, ingress.ObjectMeta.Name, metav1.GetOptions{})
-			if err != nil {
-				return err
-			}
-			ingress.Spec.IngressClassName = ingressClassName
-			_, err = ingressClient.Update(ctx, ingress, metav1.UpdateOptions{})
+func setIngressClassNameWithRetry(ctx context.Context, namespace string, ingress *netv1.Ingress, ingressClassName *string) error {
+	ingressClient := env.Cluster().Client().NetworkingV1().Ingresses(namespace)
+	return retry.RetryOnConflict(retry.DefaultRetry, func() error {
+		ingress, err := ingressClient.Get(ctx, ingress.Name, metav1.GetOptions{})
+		if err != nil {
 			return err
-		})
-	case *netv1beta1.Ingress:
-		ingressClient := env.Cluster().Client().NetworkingV1beta1().Ingresses(namespace)
-		return retry.RetryOnConflict(retry.DefaultRetry, func() error {
-			ingress, err := ingressClient.Get(ctx, ingress.ObjectMeta.Name, metav1.GetOptions{})
-			if err != nil {
-				return err
-			}
-			ingress.Spec.IngressClassName = ingressClassName
-			_, err = ingressClient.Update(ctx, ingress, metav1.UpdateOptions{})
-			return err
-		})
-	}
-	return fmt.Errorf("unsupported GroupVersionKind %v", obj.GetObjectKind())
+		}
+		ingress.Spec.IngressClassName = ingressClassName
+		_, err = ingressClient.Update(ctx, ingress, metav1.UpdateOptions{})
+		return err
+	})
 }
 
 // Expression router is not supported for some objects and features.